requirements.txt
ruff
black
aws-cdk-lib==2.66.1
constructs>=10.0.0,<11.0.0
boto3~=1.26.66
botocore~=1.29.66
